Refactor client types and options
