+++ packages/client/src/types/unified.ts
@@ -550,6 +550,8 @@ export type MsgsResponse = {
 
 export type BridgeType = 'IBC' | 'AXELAR' | 'CCTP' | 'HYPERLANE' | 'OPINIT';
 
+export type ChainType = 'cosmos' | 'evm' | 'svm';
+
 export type AssetBetweenChainsJSON = {
   asset_on_source: AssetJSON;
   asset_on_dest: AssetJSON;
